refactor(discord): split placeholder suppression module (#4176)
